requirements.txt
anthropic==0.30.1
beautifulsoup4==4.12.3
docx==0.2.4
matplotlib==3.9.1
numpy==2.0.0
openai==1.35.10
pandas==2.2.2
python-dotenv==1.0.1
python_docx==1.1.2
requests==2.31.0
streamlit==1.34.0
